Refactor all the danger ore maps
