+++ map_gen/maps/danger_ores/config/bob_ores.lua
@@ -1,33 +1,7 @@
 local b = require 'map_gen.shared.builders'
-local Perlin = require 'map_gen.shared.perlin_noise'
-local table = require 'utils.table'
-
-local random = math.random
-local floor = math.floor
 local value = b.euclidean_value
-local binary_search = table.binary_search
-local bnot = bit32.bnot
-local perlin_noise = Perlin.noise
 
-local mixed_ores = true
-
-local tile_scale = 1 / 64
-local spawn_zone = b.circle(85)
-
-local oil_scale = 1 / 64
-local oil_threshold = 0.6
-
-local uranium_scale = 1 / 72
-local uranium_threshold = 0.63
-
-local thorium_scale = 1 / 72
-local thorium_threshold = 0.63
-
-local density_scale = 1 / 48
-local density_threshold = 0.5
-local density_multiplier = 50
-
-local ores = {
+return {
     ['iron-ore'] = {
         ['tiles'] = {
             [1] = 'grass-1',
@@ -36,7 +10,6 @@ local ores = {
             [4] = 'grass-4'
         },
         ['start'] = value(125, 0),
-        ['non_mixed_value'] = value(0, 0.5),
         ['weight'] = 15,
         ['ratios'] = {
             {resource = b.resource(b.full_shape, 'iron-ore', value(0, 0.5)), weight = 120},
@@ -66,7 +39,6 @@ local ores = {
             [4] = 'red-desert-3'
         },
         ['start'] = value(125, 0),
-        ['non_mixed_value'] = value(0, 0.5),
         ['weight'] = 10,
         ['ratios'] = {
             {resource = b.resource(b.full_shape, 'iron-ore', value(0, 0.5)), weight = 20},
@@ -99,7 +71,6 @@ local ores = {
             [7] = 'dirt-7'
         },
         ['start'] = value(125, 0),
-        ['non_mixed_value'] = value(0, 0.5),
         ['weight'] = 10,
         ['ratios'] = {
             {resource = b.resource(b.full_shape, 'iron-ore', value(0, 0.5)), weight = 20},
@@ -128,7 +99,6 @@ local ores = {
             [3] = 'sand-3'
         },
         ['start'] = value(125, 0),
-        ['non_mixed_value'] = value(0, 0.5),
         ['weight'] = 10,
         ['ratios'] = {
             {resource = b.resource(b.full_shape, 'iron-ore', value(0, 0.5)), weight = 20},
@@ -158,7 +128,6 @@ local ores = {
             [4] = 'grass-4'
         },
         ['start'] = value(125, 0),
-        ['non_mixed_value'] = value(0, 0.5),
         ['weight'] = 2,
         ['ratios'] = {
             {resource = b.resource(b.full_shape, 'iron-ore', value(0, 0.5)), weight = 20},
@@ -188,7 +157,6 @@ local ores = {
             [4] = 'red-desert-3'
         },
         ['start'] = value(125, 0),
-        ['non_mixed_value'] = value(0, 0.5),
         ['weight'] = 0.05,
         ['ratios'] = {
             {resource = b.resource(b.full_shape, 'iron-ore', value(0, 0.5)), weight = 20},
@@ -221,7 +189,6 @@ local ores = {
             [7] = 'dirt-7'
         },
         ['start'] = value(125, 0),
-        ['non_mixed_value'] = value(0, 0.5),
         ['weight'] = 0.05,
         ['ratios'] = {
             {resource = b.resource(b.full_shape, 'iron-ore', value(0, 0.5)), weight = 20},
@@ -250,7 +217,6 @@ local ores = {
             [3] = 'sand-3'
         },
         ['start'] = value(125, 0),
-        ['non_mixed_value'] = value(0, 0.5),
         ['weight'] = 1,
         ['ratios'] = {
             {resource = b.resource(b.full_shape, 'iron-ore', value(0, 0.5)), weight = 20},
@@ -280,7 +246,6 @@ local ores = {
             [4] = 'grass-4'
         },
         ['start'] = value(125, 0),
-        ['non_mixed_value'] = value(0, 0.5),
         ['weight'] = 2,
         ['ratios'] = {
             {resource = b.resource(b.full_shape, 'iron-ore', value(0, 0.5)), weight = 20},
@@ -310,7 +275,6 @@ local ores = {
             [4] = 'red-desert-3'
         },
         ['start'] = value(125, 0),
-        ['non_mixed_value'] = value(0, 0.5),
         ['weight'] = 2,
         ['ratios'] = {
             {resource = b.resource(b.full_shape, 'iron-ore', value(0, 0.5)), weight = 20},
@@ -343,7 +307,6 @@ local ores = {
             [7] = 'dirt-7'
         },
         ['start'] = value(125, 0),
-        ['non_mixed_value'] = value(0, 0.5),
         ['weight'] = 2,
         ['ratios'] = {
             {resource = b.resource(b.full_shape, 'iron-ore', value(0, 0.5)), weight = 20},
@@ -372,7 +335,6 @@ local ores = {
             [3] = 'sand-3'
         },
         ['start'] = value(125, 0),
-        ['non_mixed_value'] = value(0, 0.5),
         ['weight'] = 1,
         ['ratios'] = {
             {resource = b.resource(b.full_shape, 'iron-ore', value(0, 0.5)), weight = 20},
@@ -402,7 +364,6 @@ local ores = {
             [4] = 'grass-4'
         },
         ['start'] = value(125, 0),
-        ['non_mixed_value'] = value(0, 0.5),
         ['weight'] = 1,
         ['ratios'] = {
             {resource = b.resource(b.full_shape, 'iron-ore', value(0, 0.5)), weight = 20},
@@ -432,7 +393,6 @@ local ores = {
             [4] = 'red-desert-3'
         },
         ['start'] = value(125, 0),
-        ['non_mixed_value'] = value(0, 0.5),
         ['weight'] = 1,
         ['ratios'] = {
             {resource = b.resource(b.full_shape, 'iron-ore', value(0, 0.5)), weight = 20},
@@ -465,7 +425,6 @@ local ores = {
             [7] = 'dirt-7'
         },
         ['start'] = value(125, 0),
-        ['non_mixed_value'] = value(0, 0.5),
         ['weight'] = 2,
         ['ratios'] = {
             {resource = b.resource(b.full_shape, 'iron-ore', value(0, 0.5)), weight = 20},
@@ -494,7 +453,6 @@ local ores = {
             [3] = 'sand-3'
         },
         ['start'] = value(125, 0),
-        ['non_mixed_value'] = value(0, 0.5),
         ['weight'] = 1,
         ['ratios'] = {
             {resource = b.resource(b.full_shape, 'iron-ore', value(0, 0.5)), weight = 20},
@@ -524,7 +482,6 @@ local ores = {
             [4] = 'grass-4'
         },
         ['start'] = value(125, 0),
-        ['non_mixed_value'] = value(0, 0.5),
         ['weight'] = 2,
         ['ratios'] = {
             {resource = b.resource(b.full_shape, 'iron-ore', value(0, 0.5)), weight = 20},
@@ -547,141 +504,3 @@ local ores = {
         }
     }
 }
-
-local oil_shape = b.throttle_world_xy(b.full_shape, 1, 7, 1, 7)
-local oil_resource = b.resource(oil_shape, 'crude-oil', value(250000, 150))
-
-local uranium_resource = b.resource(b.full_shape, 'uranium-ore', value(200, 1))
-local thorium_resource = b.resource(b.full_shape, 'thorium-ore', value(200, 1))
-
-local function init(seed)
-    local oil_seed = seed * 2
-    local uranium_seed = seed * 3
-    local density_seed = seed * 4
-    local thorium_seed = seed * 5
-
-    local function tile_builder(tiles)
-        local count = #tiles / 2
-        return function(x, y)
-            x, y = x * tile_scale, y * tile_scale
-            local v = perlin_noise(x, y, seed)
-            v = ((v + 1) * count) + 1
-            v = floor(v)
-            return tiles[v]
-        end
-    end
-
-    local function ore_builder(ore_name, amount, ratios, weighted)
-        local start_ore = b.resource(b.full_shape, ore_name, amount)
-        local total = weighted.total
-        return function(x, y, world)
-            if spawn_zone(x, y) then
-                return start_ore(x, y, world)
-            end
-
-            local oil_x, oil_y = x * oil_scale, y * oil_scale
-            local oil_noise = perlin_noise(oil_x, oil_y, oil_seed)
-            if oil_noise > oil_threshold then
-                return oil_resource(x, y, world)
-            end
-
-            local uranium_x, uranium_y = x * uranium_scale, y * uranium_scale
-            local uranium_noise = perlin_noise(uranium_x, uranium_y, uranium_seed)
-            if uranium_noise > uranium_threshold then
-                return uranium_resource(x, y, world)
-            end
-
-            local thorium_x, thorium_y = x * thorium_scale, y * thorium_scale
-            local thorium_noise = perlin_noise(thorium_x, thorium_y, thorium_seed)
-            if thorium_noise > thorium_threshold then
-                return thorium_resource(x, y, world)
-            end
-
-            local i = random() * total
-            local index = binary_search(weighted, i)
-            if index < 0 then
-                index = bnot(index)
-            end
-
-            local resource = ratios[index].resource
-
-            local entity = resource(x, y, world)
-            local density_x, density_y = x * density_scale, y * density_scale
-            local density_noise = perlin_noise(density_x, density_y, density_seed)
-
-            if density_noise > density_threshold then
-                entity.amount = entity.amount * density_multiplier
-            end
-
-            entity.enable_tree_removal = false
-
-            return entity
-        end
-    end
-
-    local function non_mixed_ore_builder(ore_name, amount)
-        local resource = b.resource(b.full_shape, ore_name, amount)
-        return function(x, y, world)
-            if spawn_zone(x, y) then
-                return resource(x, y, world)
-            end
-
-            local oil_x, oil_y = x * oil_scale, y * oil_scale
-            local oil_noise = perlin_noise(oil_x, oil_y, oil_seed)
-            if oil_noise > oil_threshold then
-                return oil_resource(x, y, world)
-            end
-
-            local uranium_x, uranium_y = x * uranium_scale, y * uranium_scale
-            local uranium_noise = perlin_noise(uranium_x, uranium_y, uranium_seed)
-            if uranium_noise > uranium_threshold then
-                return uranium_resource(x, y, world)
-            end
-
-            local thorium_x, thorium_y = x * thorium_scale, y * thorium_scale
-            local thorium_noise = perlin_noise(thorium_x, thorium_y, thorium_seed)
-            if thorium_noise > thorium_threshold then
-                return thorium_resource(x, y, world)
-            end
-
-            local entity = resource(x, y, world)
-            local density_x, density_y = x * density_scale, y * density_scale
-            local density_noise = perlin_noise(density_x, density_y, density_seed)
-
-            if density_noise > density_threshold then
-                entity.amount = entity.amount * density_multiplier
-            end
-
-            entity.enable_tree_removal = false
-
-            return entity
-        end
-    end
-
-    local shapes = {}
-
-    for ore_name, v in pairs(ores) do
-        local tiles = v.tiles
-        local land = tile_builder(tiles)
-
-        local ore
-        if mixed_ores then
-            local ratios = v.ratios
-            local weighted = b.prepare_weighted_array(ratios)
-            local amount = v.start
-
-            ore = ore_builder(ore_name, amount, ratios, weighted)
-        else
-            local amount = v.non_mixed_value
-
-            ore = non_mixed_ore_builder(ore_name, amount)
-        end
-
-        local shape = b.apply_entity(land, ore)
-        shapes[#shapes + 1] = {shape = shape, weight = v.weight}
-    end
-
-    return shapes
-end
-
-return init
