+++ map_gen/maps/danger_ores/config/bobangels_allowed_entities.lua
@@ -1,0 +1,53 @@
+return {
+    'basic-transport-belt',
+    'transport-belt',
+    'fast-transport-belt',
+    'express-transport-belt',
+    'turbo-transport-belt',
+    'ultimate-transport-belt',
+    'basic-underground-belt',
+    'underground-belt',
+    'fast-underground-belt',
+    'express-underground-belt',
+    'turbo-underground-belt',
+    'ultimate-underground-belt',
+    'small-electric-pole',
+    'medium-electric-pole',
+    'medium-electric-pole-2',
+    'medium-electric-pole-3',
+    'medium-electric-pole-4',
+    'big-electric-pole',
+    'big-electric-pole-2',
+    'big-electric-pole-3',
+    'big-electric-pole-4',
+    'substation',
+    'substation-2',
+    'substation-3',
+    'substation-4',
+    'electric-mining-drill',
+    'bob-mining-drill-1',
+    'bob-mining-drill-2',
+    'bob-mining-drill-3',
+    'bob-mining-drill-4',
+    'bob-area-mining-drill-1',
+    'bob-area-mining-drill-2',
+    'bob-area-mining-drill-3',
+    'bob-area-mining-drill-4',
+    'burner-mining-drill',
+    'pumpjack',
+    'bob-pumpjack-1',
+    'bob-pumpjack-2',
+    'bob-pumpjack-3',
+    'bob-pumpjack-4',
+    'water-miner-1',
+    'water-miner-2',
+    'water-miner-3',
+    'water-miner-4',
+    'water-miner-5',
+    'car',
+    'tank',
+    'bob-tank-2',
+    'bob-tank-3',
+    'spidertron',
+    'thermal-water-extractor'
+}
