+++ map_gen/maps/danger_ores/config/vanilla_allowed_entities.lua
@@ -1,0 +1,18 @@
+return {
+    'transport-belt',
+    'fast-transport-belt',
+    'express-transport-belt',
+    'underground-belt',
+    'fast-underground-belt',
+    'express-underground-belt',
+    'small-electric-pole',
+    'medium-electric-pole',
+    'big-electric-pole',
+    'substation',
+    'electric-mining-drill',
+    'burner-mining-drill',
+    'pumpjack',
+    'car',
+    'tank',
+    'spidertron'
+}
